moved survey sending to after summary page
